- Optimised Select rule with table lookup