@bvssvni 2018-12-05
Some notes:

- Made optimization enabled by default, because otherwise there is a lot of work to add a `.optimize()` call wherever Piston-Meta is used.
- Added lazy static optimized bootstrap rules